+++ docs/news.rst
@@ -4,6 +4,10 @@ Release Notes
 1.11.0 (unreleased)
 -------------------
 
+* Add :code:`firefox_arguments` and :code:`firefox_preferences` markers to
+  specify arguments and preferences to pass to the :code:`firefox_options`
+  fixture. Run :code:`pytest --markers` for details.
+
 * Restore host and port in HTML report when using defaults.
 
 * Warn in pytest header when the sensitive URL matches the base URL.